check role

+++ app/Http/Controllers/CategoryController.php
@@ -7,6 +7,12 @@ use App\Category;
 class CategoryController extends Controller
 {
     public function index(){
+
+        if(!(Auth()->user()->hasRole('admin')))
+        {
+            return abort(403, 'Unauthorized action.');
+        }
+
         $categories = Category::paginate('4');
         return view('admin.category.index',compact('categories'));
     }
@@ -16,6 +22,11 @@ class CategoryController extends Controller
     }
 
     public function store(Request $request){
+
+        if(!(Auth()->user()->hasRole('admin')))
+        {
+            return abort(403, 'Unauthorized action.');
+        }
 
         $file = $request->file('image');
         $name = $file->getClientOriginalName();
@@ -30,12 +41,23 @@ class CategoryController extends Controller
 
     public function edit($id){
 
+        if(!(Auth()->user()->hasRole('admin')))
+        {
+            return abort(403, 'Unauthorized action.');
+        }
+
         $categories = Category::find($id);
 
         return view('admin.category.edit',compact('categories'));
     }
 
     public function update(Request $request, $id){
+
+        if(!(Auth()->user()->hasRole('admin')))
+        {
+            return abort(403, 'Unauthorized action.');
+        }
+
         $categories = Category::find($id);
         $file = $request->file('image');
         $name = $file->getClientOriginalName();
@@ -47,6 +69,12 @@ class CategoryController extends Controller
     }
 
     public function destroy($id){
+
+        if(!(Auth()->user()->hasRole('admin')))
+        {
+            return abort(403, 'Unauthorized action.');
+        }
+
         $categories = Category::find($id);
         $categories->delete();
         return redirect()->route('list_category');
